+++ Assets/UGF.Application.Runtime.Tests/TestApplicationConfigLauncher.cs
@@ -1,0 +1,6 @@
+﻿namespace UGF.Application.Runtime.Tests
+{
+    public class TestApplicationConfigLauncher : ApplicationConfigLauncher
+    {
+    }
+}