+++ changelog.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased - 2019-01-01
+## Unreleased - 2020-01-01
 - [Commits](https://github.com/unity-game-framework/ugf-application/compare/0.0.0...0.0.0)
 - [Milestone](https://github.com/unity-game-framework/ugf-application/milestone/0?closed=1)
 
@@ -25,6 +25,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 - Nothing.
+
+## 3.2.0-preview - 2020-01-13
+- [Commits](https://github.com/unity-game-framework/ugf-application/compare/3.1.0-preview...3.2.0-preview)
+- [Milestone](https://github.com/unity-game-framework/ugf-application/milestone/9?closed=1)
+
+### Added
+- `ApplicationLauncher`: `Launched`, `Stopped` and `Quttings` events.
+- `ApplicationLauncher.UninitializeApplication` method wich invoked after launcher is stopped.
+- `ApplicationLauncherEvents` component with `Launched`, `Stopped` and `Quttings` Unity events.
 
 ## 3.1.0-preview - 2020-01-11
 - [Commits](https://github.com/unity-game-framework/ugf-application/compare/3.0.0-preview...3.1.0-preview)